Moved transaction-related modules from output package to db package
