+++ src/pds_doi_service/core/util/initialize_production_deployment.py
@@ -63,6 +63,7 @@ import logging
 import os
 from datetime import datetime
 
+from pds_doi_service.core.db.transaction_builder import TransactionBuilder
 from pds_doi_service.core.entities.exceptions import CriticalDOIException
 from pds_doi_service.core.entities.exceptions import InputFormatException
 from pds_doi_service.core.entities.exceptions import UnknownNodeException
@@ -71,7 +72,6 @@ from pds_doi_service.core.outputs.osti.osti_web_parser import DOIOstiXmlWebParse
 from pds_doi_service.core.outputs.service import DOIServiceFactory
 from pds_doi_service.core.outputs.service import SERVICE_TYPE_DATACITE
 from pds_doi_service.core.outputs.service import VALID_SERVICE_TYPES
-from pds_doi_service.core.outputs.transaction_builder import TransactionBuilder
 from pds_doi_service.core.util.config_parser import DOIConfigUtil
 from pds_doi_service.core.util.general_util import get_logger
 from pds_doi_service.core.util.node_util import NodeUtil
@@ -86,7 +86,7 @@ m_config = m_doi_config_util.get_config()
 
 def create_cmd_parser():
     parser = argparse.ArgumentParser(
-        description="Script to bulk import existing DOIs into the local " "transaction database.",
+        description="Script to bulk import existing DOIs into the local transaction database.",
         epilog="Note: When DOI records are imported to the local transaction "
         "database, the DOI service creates an associated output label "
         "for each record under the transaction_history directory. The "
@@ -258,14 +258,14 @@ def _read_from_path(service, path):
 
     """
     if not os.path.exists(path):
-        raise InputFormatException(f"Error reading file {path}. " "File may not exist.")
+        raise InputFormatException(f"Error reading file {path}. File may not exist.")
 
     if path.endswith(".xml"):
         return _read_from_local_xml(path)
     elif path.endswith(".json"):
         return _read_from_local_json(service, path)
 
-    raise InputFormatException(f"File {path} is not supported. " f"Only .xml and .json are supported.")
+    raise InputFormatException(f"File {path} is not supported. Only .xml and .json are supported.")
 
 
 def get_dois_from_provider(service, prefix, output_file=None):
@@ -391,7 +391,7 @@ def perform_import_to_database(service, prefix, db_name, input_source, dry_run, 
         # If the field 'pds_identifier' is None, we cannot proceed since
         # it serves as the primary key for our transaction database.
         if not doi.pds_identifier:
-            logger.warning("Skipping DOI with missing PDS identifier %s, " "index %d", doi.doi, item_index)
+            logger.warning("Skipping DOI with missing PDS identifier %s, index %d", doi.doi, item_index)
 
             o_records_dois_skipped += 1
             continue
